+++ changelog.md
@@ -1,6 +1,21 @@
+# v1.0.2
+ * Progression data is now reset when leaving from the minigame
+  * Previous seed is also being reset when changing to newer one if this behaviour is enabled.
+  * Change behaviour using "Reset Previous Seed" option.
+ * Fixed cursor being shown while playing levels
+ * Fixed two possible crashes due to UB
+ * Fixed crash when changing seed
+ * Fixed usability issues when chaning seed
+ * Fixed music information synchronization
+ * Fixed issues with playing music after beating a level
+ * Improved FMOD manipulation
+ * Skipping levels will increase next level difficulty if this behaviour is enabled
+  * Change behaviour using "Update Difficulty After Skip" option.
+ * Added debug mode
 # v1.0.1
  * Bugfixes
  * Levels can be skipped now
   * Pressing the Skip button costs 10 lives by default.
+ * Added changelog
 # v1.0.0
- * Initial relase
+ * Initial release